+++ cli/main.rs
@@ -3,6 +3,7 @@
 mod args;
 mod auth_tokens;
 mod cache;
+mod cdp;
 mod deno_std;
 mod emit;
 mod errors;
